feat: Minor refactoring
