+++ pki-application/src/main/java/com/example/pki/pkiapplication/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.example.pki.pkiapplication.repository;
+
+import com.example.pki.pkiapplication.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
