+++ pki-application/src/main/java/com/example/pki/pkiapplication/service/CSRService.java
@@ -1,7 +1,6 @@
 package com.example.pki.pkiapplication.service;
 
 import com.example.pki.pkiapplication.model.CSR;
-import com.example.pki.pkiapplication.service.impl.JpaService;
 
 public interface CSRService extends JpaService<CSR> {
