+++ pki-application/src/main/java/com/example/pki/pkiapplication/service/CertificateService.java
@@ -1,7 +1,6 @@
 package com.example.pki.pkiapplication.service;
 
 import com.example.pki.pkiapplication.model.Certificate;
-import com.example.pki.pkiapplication.service.impl.JpaService;
 
 public interface CertificateService extends JpaService<Certificate> {
