+++ pki-application/src/main/java/com/example/pki/pkiapplication/service/impl/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
-package com.example.pki.pkiapplication.service;
+package com.example.pki.pkiapplication.service.impl;
 
 import com.example.pki.pkiapplication.model.User;
-import com.example.pki.pkiapplication.repository.IUserRepository;
+import com.example.pki.pkiapplication.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -13,7 +13,7 @@ import java.util.Optional;
 public class UserDetailsServiceImpl implements UserDetailsService {
 
     @Autowired
-    private IUserRepository userRepository;
+    private UserRepository userRepository;
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
         Optional<User> ret = userRepository.findByEmail(email);
